+++ libraries/wallet/CMakeLists.txt
@@ -29,9 +29,7 @@ else()
                       DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/api_documentation_standin.cpp )
 endif()
 
-file(GLOB HEADERS "include/graphene/wallet/*.hpp")
-
-file (GLOB SOURCES "*.cpp" ${HEADERS})
+file (GLOB SOURCES "*.cpp")
 
 add_library( graphene_wallet ${SOURCES} ${CMAKE_CURRENT_BINARY_DIR}/api_documentation.cpp ${HEADERS} )
 target_link_libraries( graphene_wallet PRIVATE graphene_app graphene_net graphene_chain graphene_utilities fc ${CMAKE_DL_LIBS} ${PLATFORM_SPECIFIC_LIBS} )
